refactor: split typechecker methods into separate files
